Use a single SGList format internally based on __PADDR_NBITS__
We previously had 2 sglist formats used inside of the kernel. This is
not necessary since the kernel's internal format doesn't need to be
what is returned to userspace, so we can use a single sglist format
in the kernel, per-build, based on the paddrspace size.

This eliminates the need for all the if(){}else{}-ing based on the
addrSize argument.